requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Keras==2.6.0
numpy==1.21.5
Pillow==8.4.0
streamlit==1.18.0
tensorflow==2.13.0